requirements.txt
pydantic>=2.0
pyyaml
tabula-py
jpype1
